requirements.txt
django>=4.2.0
google-cloud-dialogflow>=2.20.0
python-dotenv>=1.0.0
google-auth>=2.22.0
google-cloud-storage>=2.10.0
requests>=2.31.0 